moved type definitions out of traits.rs
